refactor: remove old payload collator
