+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/session/message/PayloadFactoryImpl.kt
@@ -9,17 +9,10 @@ import io.embrace.android.embracesdk.session.lifecycle.ProcessState
 import io.embrace.android.embracesdk.session.orchestrator.SessionSnapshotType
 
 internal class PayloadFactoryImpl(
-    private val v1payloadMessageCollator: PayloadMessageCollator,
-    private val v2payloadMessageCollator: PayloadMessageCollator,
+    private val payloadMessageCollator: PayloadMessageCollator,
     private val configService: ConfigService,
     private val logger: EmbLogger
 ) : PayloadFactory {
-
-    private val collator: PayloadMessageCollator
-        get() = when {
-            configService.oTelBehavior.isDevEnabled() -> v2payloadMessageCollator
-            else -> v1payloadMessageCollator
-        }
 
     override fun startPayloadWithState(state: ProcessState, timestamp: Long, coldStart: Boolean) =
         when (state) {
@@ -50,7 +43,7 @@ internal class PayloadFactoryImpl(
         }
 
     override fun startSessionWithManual(timestamp: Long): Session {
-        return collator.buildInitialSession(
+        return payloadMessageCollator.buildInitialSession(
             InitialEnvelopeParams.SessionParams(
                 false,
                 LifeEventType.MANUAL,
@@ -60,7 +53,7 @@ internal class PayloadFactoryImpl(
     }
 
     override fun endSessionWithManual(timestamp: Long, initial: Session): SessionMessage {
-        return collator.buildFinalSessionMessage(
+        return payloadMessageCollator.buildFinalSessionMessage(
             FinalEnvelopeParams.SessionParams(
                 initial = initial,
                 endTime = timestamp,
@@ -72,7 +65,7 @@ internal class PayloadFactoryImpl(
     }
 
     private fun startSessionWithState(timestamp: Long, coldStart: Boolean): Session {
-        return collator.buildInitialSession(
+        return payloadMessageCollator.buildInitialSession(
             InitialEnvelopeParams.SessionParams(
                 coldStart,
                 LifeEventType.STATE,
@@ -92,7 +85,7 @@ internal class PayloadFactoryImpl(
             coldStart -> timestamp
             else -> timestamp + 1
         }
-        return collator.buildInitialSession(
+        return payloadMessageCollator.buildInitialSession(
             InitialEnvelopeParams.BackgroundActivityParams(
                 coldStart = coldStart,
                 startType = LifeEventType.BKGND_STATE,
@@ -102,7 +95,7 @@ internal class PayloadFactoryImpl(
     }
 
     private fun endSessionWithState(initial: Session, timestamp: Long): SessionMessage {
-        return collator.buildFinalSessionMessage(
+        return payloadMessageCollator.buildFinalSessionMessage(
             FinalEnvelopeParams.SessionParams(
                 initial = initial,
                 endTime = timestamp,
@@ -120,7 +113,7 @@ internal class PayloadFactoryImpl(
 
         // kept for backwards compat. the backend expects the start time to be 1 ms greater
         // than the adjacent session, and manually adjusts.
-        return collator.buildFinalBackgroundActivityMessage(
+        return payloadMessageCollator.buildFinalBackgroundActivityMessage(
             FinalEnvelopeParams.BackgroundActivityParams(
                 initial = initial,
                 endTime = timestamp - 1,
@@ -136,7 +129,7 @@ internal class PayloadFactoryImpl(
         timestamp: Long,
         crashId: String
     ): SessionMessage {
-        return collator.buildFinalSessionMessage(
+        return payloadMessageCollator.buildFinalSessionMessage(
             FinalEnvelopeParams.SessionParams(
                 initial = initial,
                 endTime = timestamp,
@@ -156,7 +149,7 @@ internal class PayloadFactoryImpl(
         if (!configService.isBackgroundActivityCaptureEnabled()) {
             return null
         }
-        return collator.buildFinalBackgroundActivityMessage(
+        return payloadMessageCollator.buildFinalBackgroundActivityMessage(
             FinalEnvelopeParams.BackgroundActivityParams(
                 initial = initial,
                 endTime = timestamp,
@@ -172,7 +165,7 @@ internal class PayloadFactoryImpl(
      * Called when the session is persisted every 2s to cache its state.
      */
     private fun snapshotSession(initial: Session, timestamp: Long): SessionMessage {
-        return collator.buildFinalSessionMessage(
+        return payloadMessageCollator.buildFinalSessionMessage(
             FinalEnvelopeParams.SessionParams(
                 initial = initial,
                 endTime = timestamp,
@@ -187,7 +180,7 @@ internal class PayloadFactoryImpl(
         if (!configService.isBackgroundActivityCaptureEnabled()) {
             return null
         }
-        return collator.buildFinalBackgroundActivityMessage(
+        return payloadMessageCollator.buildFinalBackgroundActivityMessage(
             FinalEnvelopeParams.BackgroundActivityParams(
                 initial = initial,
                 endTime = timestamp,
