+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/injection/FakeSessionModule.kt
@@ -10,8 +10,7 @@ import io.embrace.android.embracesdk.logging.EmbLoggerImpl
 import io.embrace.android.embracesdk.session.caching.PeriodicBackgroundActivityCacher
 import io.embrace.android.embracesdk.session.caching.PeriodicSessionCacher
 import io.embrace.android.embracesdk.session.message.PayloadFactory
-import io.embrace.android.embracesdk.session.message.V1PayloadMessageCollator
-import io.embrace.android.embracesdk.session.message.V2PayloadMessageCollator
+import io.embrace.android.embracesdk.session.message.PayloadMessageCollatorImpl
 import io.embrace.android.embracesdk.session.orchestrator.SessionOrchestrator
 import io.embrace.android.embracesdk.session.properties.SessionPropertiesService
 
@@ -21,10 +20,7 @@ internal class FakeSessionModule(
     override val sessionOrchestrator: SessionOrchestrator = FakeSessionOrchestrator()
 ) : SessionModule {
 
-    override val v1PayloadMessageCollator: V1PayloadMessageCollator
-        get() = TODO("Not yet implemented")
-
-    override val v2PayloadMessageCollator: V2PayloadMessageCollator
+    override val payloadMessageCollatorImpl: PayloadMessageCollatorImpl
         get() = TODO("Not yet implemented")
 
     override val periodicSessionCacher: PeriodicSessionCacher
