--- embrace-android-sdk/src/test/resources/v1_session_expected.json
@@ -1,25 +1,0 @@
-{
-  "s": {
-    "as": "__EMBRACE_TEST_IGNORE__",
-    "bf": "__EMBRACE_TEST_IGNORE__",
-    "cs": true,
-    "et": "__EMBRACE_TEST_IGNORE__",
-    "em": "s",
-    "ce": true,
-    "el": [],
-    "ss": "__EMBRACE_TEST_IGNORE__",
-    "il": [],
-    "ht": "__EMBRACE_TEST_IGNORE__",
-    "sn": "__EMBRACE_TEST_IGNORE__",
-    "sp": {},
-    "si": "__EMBRACE_TEST_IGNORE__",
-    "id": "__EMBRACE_TEST_IGNORE__",
-    "st": "__EMBRACE_TEST_IGNORE__",
-    "sm": "s",
-    "sd": "__EMBRACE_TEST_IGNORE__",
-    "sdt": "__EMBRACE_TEST_IGNORE__",
-    "nc": [],
-    "wl": [],
-    "e": "__EMBRACE_TEST_IGNORE__"
-  }
-}
